// ANGULAR/prueba/src/app/modelos/listausuarios.interface.ts
export interface ListaUsuariosI{
    $id?:             string;
    id?:              number;
    usuario?:         string;
    primerNombre?:    string;
    segundoNombre?:   string;
    primerApellido?:  string;
    segundoApellido?: string;
    idDepartamento?:  number;
    idCargo?:         number;
    departamento?:    string;
    cargo?:           string;
    
}